* subversion/bindings/javahl/native/DiffOptions.cpp (get_diff_options_flags):    Meh. Use the right type of field accessor.
git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1456336 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/bindings/javahl/native/DiffOptions.cpp
@@ -42,7 +42,7 @@ static jint get_diff_options_flags(jobject joptions)
         return 0;
     }
 
-  const jint flags = env->GetLongField(joptions, fid);
+  const jint flags = env->GetIntField(joptions, fid);
   if (JNIUtil::isJavaExceptionThrown())
     return 0;
   return flags;
